pyproject.toml
[tool.poetry]
name = "airflow-tutorial"
version = "0.1.0"
description = ""
authors = ["kimdonghyeon <kimdong799@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "^2.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
